pyproject.toml
[project]
name = "ai-content-factory"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Mariam-Srour2003", email = "mariammsrourr2020@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    # Core AI/ML frameworks
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-core>=0.3.0",
    # LLM serving
    "ollama>=0.3.0",
    # Vector database
    "chromadb>=0.5.0",
    # Data validation
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    # Web scraping
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "lxml>=5.0.0",
    # NLP tools
    "spacy>=3.7.0",
    "nltk>=3.8.0",
    # Text processing
    "markdown>=3.5.0",
    "python-dotenv>=1.0.0",
    # Embeddings
    "sentence-transformers>=2.7.0",
    # UI
    "gradio>=4.0.0",
    # Utilities
    "python-slugify>=8.0.0",
    "pyyaml>=6.0.0",
    "aiohttp>=3.9.0",
    "langchain-chroma>=1.0.0",
    "langchain-ollama>=1.0.0",
    "faiss-cpu",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
    "ipython>=8.20.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
    "ipython>=8.20.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
